# secureprompt-schemas/openapi/v1/openapi.yaml
openapi: 3.0.3
info:
  title: SecurePrompt Gateway API
  version: "1.0.0"
  description: >
    SecurePrompt is an LLM security gateway that sits between enterprise
    applications and AI providers. It redacts PII/secrets, enforces policy,
    tracks token costs, and provides governance analytics.

    Two auth schemes are used:
    - **JWT** (`Authorization: Bearer eyJ...`) — dashboard routes
    - **API Key** (`Authorization: Bearer sp-...`) — gateway + MCP utility routes
servers:
  - url: http://localhost:8080
    description: Local dev Rust gateway
  - url: https://api.secureprompt.tech
    description: Cloud
...
paths:
...
  /v1/policy-rules:
    get:
      tags: [PolicyRules]
      summary: List policy rules
      operationId: listPolicyRules
      security:
        - bearerAuth: []
      responses:
        "200":
          description: List of policy rules
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PolicyRuleResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
    post:
      tags: [PolicyRules]
      summary: Create a policy rule (admin only)
      description: >
        Action must be one of: `deny`, `allow`, `redact`, `transform`, `flag`.
        Priority must be unique within the workspace (409 on conflict).
        The rule is immediately indexed in Qdrant for semantic policy lookup.
      operationId: createPolicyRule
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateRuleRequest"
      responses:
        "201":
          description: Rule created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PolicyRuleResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "409":
          description: Priority already in use
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PriorityConflict"
  /v1/policy-rules/{id}:
    get:
      tags: [PolicyRules]
      summary: Get a single policy rule
      operationId: getPolicyRule
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/IdPath"
      responses:
        "200":
          description: Policy rule
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PolicyRuleResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
    put:
      tags: [PolicyRules]
      summary: Update a policy rule (admin only)
      operationId: updatePolicyRule
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/IdPath"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateRuleRequest"
      responses:
        "200":
          description: Rule updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PolicyRuleResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          description: Priority already in use
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PriorityConflict"
    delete:
      tags: [PolicyRules]
      summary: Delete a policy rule (admin only)
      operationId: deletePolicyRule
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/IdPath"
      responses:
        "204":
          description: Rule deleted
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
  /v1/policy-rules/{id}/enabled:
    patch:
      tags: [PolicyRules]
      summary: Toggle enabled flag on a policy rule (admin only)
      operationId: togglePolicyRuleEnabled
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/IdPath"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ToggleRequest"
      responses:
        "200":
          description: Rule updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PolicyRuleResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
  /v1/policy-rules/{id}/dry-run:
    patch:
      tags: [PolicyRules]
      summary: Toggle dry-run flag on a policy rule (admin only)
      operationId: togglePolicyRuleDryRun
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/IdPath"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ToggleRequest"
      responses:
        "200":
          description: Rule updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PolicyRuleResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"

  # ── Workspace Budgets ─────────────────────────────────────────────────────────
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: "Dashboard JWT — Authorization: Bearer eyJ..."
    apiKeyAuth:
      type: http
      scheme: bearer
      description: "API key — Authorization: Bearer sp-..."
  parameters:
    IdPath:
      in: path
      name: id
      required: true
      schema:
        type: string
        format: uuid
...
  responses:
    BadRequest:
      description: Bad request — validation failed
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ApiError"
    Unauthorized:
      description: Missing or invalid bearer token
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ApiError"
    Forbidden:
      description: Insufficient role or workspace mismatch
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ApiError"
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ApiError"
...
  schemas:
...
    PolicyAction:
      type: string
      enum: [deny, allow, redact, transform, flag]
    PolicyRuleResponse:
      type: object
      required:
        - id
        - workspace_id
        - name
        - priority
        - conditions
        - action
        - action_params
        - enabled
        - dry_run
        - created_at
        - updated_at
      properties:
        id:
          type: string
          format: uuid
        workspace_id:
          type: string
          format: uuid
        name:
          type: string
        priority:
          type: integer
          description: Lower = higher precedence; must be unique per workspace
        conditions:
          type: object
          description: Rule conditions (arbitrary JSON)
        action:
          $ref: "#/components/schemas/PolicyAction"
        action_params:
          type: object
          description: Action parameters (arbitrary JSON)
        enabled:
          type: boolean
        dry_run:
          type: boolean
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    CreateRuleRequest:
      type: object
      required: [name, priority, action]
      properties:
        name:
          type: string
        priority:
          type: integer
        conditions:
          type: object
          nullable: true
        action:
          $ref: "#/components/schemas/PolicyAction"
        action_params:
          type: object
          nullable: true
        enabled:
          type: boolean
          default: true
        dry_run:
          type: boolean
          default: false
    UpdateRuleRequest:
      type: object
      required: [name, priority, action]
      properties:
        name:
          type: string
        priority:
          type: integer
        conditions:
          type: object
          nullable: true
        action:
          $ref: "#/components/schemas/PolicyAction"
        action_params:
          type: object
          nullable: true
        enabled:
          type: boolean
          default: true
        dry_run:
          type: boolean
          default: false
    ToggleRequest:
      type: object
      required: [value]
      properties:
        value:
          type: boolean
    PriorityConflict:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
          example: priority_conflict
        message:
          type: string

    # ── Telemetry ───────────────────────────────────────────────────────────────
...
    ApiError:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [message, type]
          properties:
            message:
              type: string
            type:
              type: string
              example: secureprompt_error
            code:
              type: string
              description: Stable machine-readable code (e.g. budget_exceeded)
...